AdditionOfSomeValidations

+++ Src/Services/Eshop.Orders/Services/CartService.cs
@@ -13,12 +13,15 @@ namespace Eshop.Orders.Services
     {
         private readonly OrderDbContext _context;
         private readonly IHttpContextAccessor _httpContextAccessor;
+        private readonly IRequestClient<VerifyProductExistence> _client;
+        private readonly IRequestClient<ProductStockRequest> _stockClient;
 
-        public CartService(OrderDbContext context, IHttpContextAccessor httpContextAccessor)
+        public CartService(OrderDbContext context, IHttpContextAccessor httpContextAccessor, IRequestClient<VerifyProductExistence> client, IRequestClient<ProductStockRequest> stockClient)
         {
             _context = context;
             _httpContextAccessor = httpContextAccessor;
-           
+            _client = client;
+            _stockClient = stockClient;
         }
         private string? GetUserId()
         {
@@ -28,10 +31,26 @@ namespace Eshop.Orders.Services
 
         public async Task<CartItem> AddCartItem(CartItemDto cartItem)
         {
-            if (cartItem == null)
+            if (cartItem == null || cartItem.Quantity <= 0 || cartItem.FullPrice <= 0)
             {
                 return null;
             }
+            var ProductExist= await _client.GetResponse<ProductExistenceResponse>(new VerifyProductExistence( cartItem.ProductId));
+
+            if(ProductExist.Message.Exists == false)
+            {
+                return null;
+            }
+
+            var StockAvailable = await _stockClient.GetResponse<ProductStockResponse>(new ProductStockRequest(cartItem.ProductId,cartItem.Quantity));
+
+            if (StockAvailable.Message.HasEnoguhStock == false)
+            {
+                return null;
+            }
+
+            var userCart =await _context.Carts.Include(i=>i.CartItems).Where(u=>u.UserId == GetUserId()).FirstOrDefaultAsync();
+
             var cartItemEntity = new CartItem()
             {
                 ProductId = cartItem.ProductId,
@@ -40,37 +59,39 @@ namespace Eshop.Orders.Services
                 FullPrice = cartItem.FullPrice,
                 CartId = cartItem.CartId
             };
-            if (cartItemEntity.CartId != 0)
-            {
-                var IsValidCartForUser=_context.Carts.Any(i=>i.Id== cartItemEntity.CartId && i.UserId== GetUserId());
-                if (!IsValidCartForUser)
-                {
-                    return null;
-                }
-                _context.CartItems.Add(cartItemEntity);
-            }
-            else
+            if (userCart == null)
             {
                 var cart = new Cart()
                 {
                     CartItems = new List<CartItem>() { cartItemEntity },
-                    UserId= GetUserId()
+                    UserId = GetUserId()
                 };
                 _context.Carts.Add(cart);
-
             }
-            if(await _context.SaveChangesAsync()<=0)
+            if (userCart.CartItems.Any(i => i.ProductId == cartItem.ProductId))
+            {
+                userCart.CartItems.FirstOrDefault(i => i.ProductId == cartItemEntity.ProductId).Quantity += cartItemEntity.Quantity;
+            }
+            else
+            {
+                _context.CartItems.Add(cartItemEntity);
+            }
+
+            if(await _context.SaveChangesAsync() <= 0)
             {
                 return null;
             }
-            return cartItemEntity;
+            return userCart.CartItems.FirstOrDefault(i => i.ProductId == cartItemEntity.ProductId);
         }
 
         public async Task<bool> ClearCart(int cartId)
         {
             var cart = await _context.Carts
                 .FirstOrDefaultAsync(i => i.Id==cartId && i.UserId== GetUserId());
-
+            if (cart == null)
+            {
+                return true;
+            }
             _context.Remove(cart);
 
             if (await _context.SaveChangesAsync() <= 0)
@@ -92,7 +113,10 @@ namespace Eshop.Orders.Services
             var cartItem= await _context.CartItems
                 .AsNoTracking()
                 .FirstOrDefaultAsync(ci => ci.Id == cartItemId && ci.Cart.UserId== GetUserId());
-
+            if(cartItem == null)
+            {
+                return true;
+            }
             _context.CartItems.Remove(cartItem);
 
             if(await _context.SaveChangesAsync() <= 0)
@@ -131,7 +155,7 @@ namespace Eshop.Orders.Services
             {
                 return null;
             }
-            var CartItem = await _context.CartItems.FirstOrDefaultAsync(ci => ci.ProductId ==cartItem.ProductId);
+            var CartItem = await _context.CartItems.FirstOrDefaultAsync(ci => ci.ProductId ==cartItem.ProductId && ci.Cart.UserId==GetUserId());
             CartItem.ProductName = cartItem.ProductName ?? CartItem.ProductName;
             CartItem.Quantity = cartItem.Quantity ?? CartItem.Quantity;
             CartItem.FullPrice = cartItem.FullPrice ?? CartItem.FullPrice;
